可能的开销异常

+++ hamjinRemover_dy/customize.sh
@@ -11,14 +11,16 @@ Hanjty"
     LIST_PATH="/data/adb/modules"
     for loop1 in $fuck_string_list; do
         for loop2 in $(ls -1 $LIST_PATH); do
-            if [ "$(grep "$loop1" <$LIST_PATH/"$loop2"/module.prop)" != "" ]; then
-                echo "[$(date '+%Y-%m-%d %H:%M:%S')][WARNING]: Conflict detected, uninstalled for you: $loop1."
-                chattr -R -i $LIST_PATH/"$loop2"  2>/dev/null
-                [ -f $LIST_PATH/"$i"/uninstall.sh ] && /data/adb/magisk/busybox "$LIST_PATH/$loop2/uninstall.sh"  2>/dev/null
-                rm -rf ${LIST_PATH:?}/"$loop2"
-            elif [ ! -f $LIST_PATH/"$loop2"/service.sh.rlast.sh ]; then
-                mv $LIST_PATH/"$loop2"/service.sh $LIST_PATH/"$loop2"/service.sh.rlast.sh
-                cp "$MODPATH"/patchfile/service.sh $LIST_PATH/"$loop2"/service.sh
+            if [ "$LIST_PATH/$loop2" != "/data/adb/modules/hamjinRemover" ]; then
+                if [ "$(grep "$loop1" <$LIST_PATH/"$loop2"/module.prop)" != "" ]; then
+                    echo "[$(date '+%Y-%m-%d %H:%M:%S')][WARNING]: Conflict detected, uninstalled for you: $loop1."
+                    chattr -R -i $LIST_PATH/"$loop2"  2>/dev/null
+                    sh "$LIST_PATH/$loop2/uninstall.sh" 2>/dev/null
+                    rm -rf ${LIST_PATH:?}/"$loop2"
+                elif [ -f $LIST_PATH/"$loop2"/service.sh ]; then
+                    [ ! -f $LIST_PATH/"$loop2"/service.sh.rlast.sh ] && mv $LIST_PATH/"$loop2"/service.sh $LIST_PATH/"$loop2"/service.sh.rlast.sh
+                    cp "$MODPATH"/patchfile/service.sh $LIST_PATH/"$loop2"/service.sh
+                fi
             fi
         done
     done
